refactor: vcs related funcs to root/vcs.c
