+++ winbuild/Makefile
@@ -98,6 +98,7 @@ SRCS=\
 	root/notifythread.c       \
 	root/stat.c       \
 	root/symlink.c       \
+	root/vcs.c       \
 	state.c      \
 	string.c     \
 	time.c
